requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
faiss-cpu==1.6.5
fastapi==0.88.0
filelock==3.8.0
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
keybert==0.6.0
nltk==3.7
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
sniffio==1.3.0
starlette==0.22.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
